www/requirements.txt
django-admin-tools==0.5.1
pytils==0.2
django-robots==0.9
sorl-thumbnail==11.12
django-annoying==0.7.9
south==0.8.4
#django-cleanup==0.1
django_simple_captcha==0.4.2
django-widget-tweaks==1.3
django-embed-video==0.9
django-forums==0.1.4